Two distinct updates:
1. Now we're using the "gosu_inbetween.fits" dataset of subvariables
that might be periodic.
2. I have added the "start" keyword to calculate_stuff and glue_stuff,
which allows you to start the variable analysis at a nonzero point (i.e.
if there was a horrible crash 3/4 of the way in). This is useful solely
because of how we save our progress incrementally as we go along, which I
think was a great implementation idea!