Migrating to "minusEasties" by adding a location-dependent Case to autovar.
The same "new" criterion is used (in a binary logic way) in a bunch of places, so that it's easy to change. (it's called `case3`)

I have checked to make sure that this doesn't change the number of autovarables that are periodic (it remains at 354, before and after the addition of case3).