Pointing all data paths to the files' Dropbox locations.
Some of these files haven't synced on the Dropbox yet, so this won't work until they do (which should happen as soon as zeppelin and macbook have both been connected to the internet for more than a minute or two). I'll probably add an amended commit note when I've confirmed that this works?